Fix for JIRA WB-3336
git-svn-id: https://anonsvn.ncbi.nlm.nih.gov/repos/v1/trunk/c++@85454 78c7ea69-d796-4a43-9a09-de51944f1b03

+++ include/objtools/align_format/showdefline.hpp
@@ -404,7 +404,7 @@ protected:
         int master_covered_length;     //total length covered by alignment
         int align_length;              //length of alignment
         int percent_coverage;          //query coverage 
-        int percent_identity;          //percent identity   
+        double percent_identity;       //percent identity   
         CConstRef<objects::CSeq_id> id;
         int blast_rank;                // "Rank" of defline.
         int hspNum;                    //hsp number
